+++ newbank/server/commands/NewAccount.java
@@ -1,6 +1,6 @@
 package newbank.server.commands;
 
-import newbank.server.Account;
+import newbank.server.accounts.Account;
 import newbank.server.customers.Customer;
 import newbank.server.customers.CustomerID;
 import newbank.server.customers.CustomerManager;
