+++ newbank/server/customers/Customer.java
@@ -1,9 +1,6 @@
 package newbank.server.customers;
 
-import newbank.server.Account;
-import java.util.List;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
+import newbank.server.accounts.Account;
 
 import java.util.ArrayList;
